pyproject.toml
[tool.poetry]
name = "dbt-datadict"
version = "0.1.1a1"
description = "Python CLI for automating the application of consistent field definitions to large multi-layered dbt projects."
authors = ["tom <tom@tasman.ai>"]
readme = "README.md"
license = "GNU GENERAL PUBLIC LICENSE"
homepage = "https://github.com/TasmanAnalytics/dbt-datadictionary"
packages = [
    { include = "datadict" },
]
exclude = [
    "tests",
    "Makefile", 
    "README.md",
    "pyproject.toml",
    "poetry.lock"
]
classifiers = [
    "Development Status :: 3 - Alpha",  
    "Programming Language :: Python :: 3.8",
    "Operating System :: OS Independent",
]

[tool.poetry.scripts]
datadict = "datadict:cli"

[tool.poetry.dependencies]
python = "^3.10"
"ruamel.yaml" = "^0.17.35"
click = ">=7.0,<8.1.4"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
